Add GITTestHelper.h to define test_repo path. Modify failing UnitTests to use new test_repo.
CocoaGit unit tests were originally written to run against the CocoaGit
repository itself.  The current problem with this approach is that
cloning the repository creates packed objects, and CocoaGit can't deal
with those yet.  Therefore, some tests are failing on cloned repositories.

To ensure that unit tests always run against a git
repository in a known, defined state, we're changing to using a set of
test repositories located in UnitTests/Resources.  All references to
these test repositories are #defined in GITTestHelper.h to further
isolate the test code from changes to the test repositories.