Adding back test_repo, with changes to UnitTests setup.
Apparently, it is not trivial to store a directory names ".git"
inside of a git repository.  For now, I renamed the .git dir
inside of the test_repo to "dot_git_dir".  Also, the UnitTests
run script renames this to .git before running tests and resets
it back to dot_git_dir afterwards.  It seems like we should just
be able to set the GIT_DIR env variable for the tests and have
this work.  I couldn't get this to work properly.