Add files for OmniGroup's OBError macros for easier NSError creation.
These files are extracted from OmniGroup's OmniBase.framework. The only
things we really need are the NSError category and the assertion helpers.
The full source code for all of their frameworks is available here:
http://www.omnigroup.com/developer/
http://www.omnigroup.com/ftp/pub/software/Source/MacOSX/Frameworks/OmniFrameworks-2008-09-09.zip

It is released under the "Omni Source License", which is essentially the MIT
license with a different name:
http://www.omnigroup.com/developer/sourcelicense/

The OBError macro does several things that are useful:
1. Checks the error pointer.
2. Automatically adds on top of an existing error, if there was one.
3. Creates the userInfo dictionary.
4. Adds the name of the source file and the line number.

This will hopefully help us to refactor some of the error handling
for cleaner, easier to maintain code.