Refactor GITRepo class to use the GITError macro.
In addition to using the GITError macro, I also tried to simplify
the logic in the methods that handle errors.  My strategy for doing this was to
change the 'if' clauses to test for exceptional cases, and then generate an
error and bail if there is a problem.  I think this puts the error-generation
code in the most obvious place, makes it easier to figure out what is
actually supposed to happen.