Attempts to start using MacRuby: grit loads.
Strings do not response to :each in MacRuby 0.3.0.
Monkey patch that and 'grit' will load.  Unfortunately, CocoaGit needs
to be compiled in GC mode in order to work with MacRuby.  I guess this
isn't the end of the world, but it's kind of annoying. MacRuby doesn't
appear to have the tools necessary to build a framework from scratch
so we'd be stuck testing the latest Xcode build.