Refactor GITObject init methods to do all error handling in the designated initializer.
Previously, we were checking for a type mismatch in initWithSha1:repo:error:.  However,
we have to load the data and type from the store, since the caller doesn't provide it (or does not know).
In contrast, GITObject -initWithSha1:type:data:repo:error knows both the desired type (parameter)
and the expected type ([self objectType]), which is defined in each subclass to return a constant.
Checking for a type mismatch here is cleaner and ensures that this check will always occur
regardless of how the caller creates the object.