Bugfix: Check for offset == NSNotFound in GITPackFileVersion2 -loadObjectWithSha1
GITIndex packOffsetForSha1:error: was recently changed to return NSNotFound (rather than 0)
as the offset if the object is not found in the pack file.  This makes sense, however
the object loading code was not modified accordingly, and was causing a crash if the sha1
could not be found in the pack index.

This changeset fixes GITPackFileVersion2 -loadObjectWithSha1:... to use the error-aware
version of GITIndex packOffsetForSha1:error: and also adds a test case for the
GITErrorObjectNotFound error exposed through the top-level accessor in GITRepo.