Add FIXME comments to GITPackIndex subclasses regarding a segfault.
After tracing the crashing found in f8ce83b6 the GITFileStore seems
to be alright. Porting those tests to GITPackStore and then NSLog
tracing through the stack showed some issues with offset ranging
which have been fixed in 35274279 but we're still getting crashes in
GITPackIndex on the line specified by the FIXME comments.