Store a cached copy of offsets in GITPackIndex*
The internal -loadOffsetsWithError: method should return an NSArray,
but it was returning an NSMutableArray, which was then retained
in the -offsets lazy loader method.

Instead, loadOffsetsWithError: should return an autoreleased NSArray
and the -offsets method should just cache a copy.  One potential problem
is that if the offsets need to be updated, there doesn't seem to be a way
to do that.