Got the very beginnings of a client class that can clone from git:// working
currently it will connect via the smallsockets class and pass the initial
handshake data - the server will respond with the refs it has and the client (currently)
will just read them and ignore them

next will be reading those refs in and generating and sending back a WANT list,
then reading in the packfile the server sends back and writing it to disk or
exploding it into individual objects.

WARNING: I fucked up the project stuff pretty well - I wanted to link the 'git-remote'
tool that calls the clone (and eventually a fetch) to the Git framework instead of
including all the source files and I am apparently not smart enough to figure out how
to force Xcode to do that, so I tell it the build directory is in '/opt' and then symlink
the Debug output files I care about there. Ugh.