Set "installation path" to @executable_path for Git.framework and SSH.dylib targets.
Also turn off warnings for 'Pointer Sign Comparison' (-fno-sign-compare) for the SSH.dylib
target.  There are ~150 pointer signedness warnings coming from the libssh2 code.
I suppose we could go through and cast everything, but I don't think this is a big issue.
Seeing the same 150 warning on every build does obscure real problems, so we ignore this for now.

Re: @executable_path
This is a special variable that gets embedded into the linker path
and directs the dynamic loader to look for the libraries in a path relative
to the executable.  This allows the git-remote binary to load Git.framework at runtime
as long as we keep the binary in the same path as the framework.

These settings are now the default for the Debug/Debug-GC targets.

I struggled with trying to set this up in the past for the other command line
utilities, but I think this is a reasonable solution for now. This is one
of the areas where Xcode sucks.