Create GITTransport class to handle basic network communication with a remote server.
Currently, this is one class designed to connect to a git service on port 9418.
I have an SSH variant in the pipeline that will 'exec' git-upload-pack/git-receive-pack
and allow communication over ssh.

Maybe this should be an abstract superclass or a the base of class cluster.
The main idea is just to route all the network operations through a 'transport' class
so that we can swap out the implementation (i.e. SmallSockets, NSStreams, CFStreams, etc..)
on the backend with miminal pain.  This is what the jgit project does, and it seemed
like a good idea.  Thanks jgit!