Add SSHSession/SSHChannel classes: simple Obj-C interface to libssh2 for client connections.
SSHSession handles user authentication, starting an ssh session, and provides
a way to open an ssh channel for communication. Currently, password and public-key
auth methods are supported, but I haven't provided a way to interactively supply
a password yet.  Test it with public-key password-less logins for now.

SSHChannel handles I/O for a session.

We only support 'exec'ing commands for now.  Shell support should also be easy,
but can't always be used for doing git+ssh anyway.  For remote repository setups like github,
servers are typically configured to have 1 user (git), and the shell for that
user is set to 'git-shell', which only allows the client to exec
'git-upload-pack' or 'git-receive-pack' to initiate a fetch or push
over ssh.  Since this is the most restrictive case, these same methods
will work as long as the remote user can login and git is installed on
the remote system.

The I/O interface is intentionally designed to be identical to BufferedSocket.
This should make it easy to create an SSHTransport class that uses SSH+git.

NOTE: this changeset also re-instates pointer sign comparisons for the SSH target
since those issues were fixed with the libssh2 1.0 upgrade.