Provide read-only access to a copy of the internal buffer.
This is useful for allowing clients to read all of the data from a remote
connection until the connection is closed.  Once the connection is closed,
the remaining data from any read operation will be in the buffer.

The BufferedSocket read operation return an NSData object of length == 0,
if the requested data could not be read, typically due to the socket being
closed, or not enough data available.  In this case, the client can access
any data that was received before this event by getting a copy of the
internal buffer.