Provide a way to create a packfile object with NSData containing packed objects.
The git protocol requires sending objects in pack files.  Pack index files
are not sent, as it is the responsibility of the receiving side to index/unpack
the contents.  The existing PackFile classes were immediately reading in
data from the 'path' and storing it.  I simply abstracted this one level
further, and made the designated initializer work with only an NSData object.
We might actually want a different class (PackStream?) to handle pack data
without an index, but this was a fast way to reuse the checksum features
of the PackFile class for checking packfile data received from a remote
server.