Preliminary support for reading packfile stream in GITTransport.
It seems like there are two ways to deal with incoming packfiles:

1. Read the whole thing, then deal with it.
2. Parse it as it comes in...
3. Read the stream into a buffer and start processing the stream
   in a new thread.

I guess 3 is where we eventually want to end up.  I envision
using NSStreams or something similar so that we can treat all I/O
(network or from file) as the same. Right now, that's overkill
and it will just make debugging the packfile parsing code more
complicated.

As it stands, GITTransport is a mess, but it can handle reading/writing
git-protocol 'packets' and reading a packfile stream and storing it in
an NSData object.  For the pack stream, it can either read it in one chunk
or on a per-object bases. We probably don't want all of this functionality
tied up in one class, but I'm not sure how to organize right now.