Refactor internal methods and add full read support for DeltaRef/DeltaOfs packed object types.
Rather than try to do everything in loadObjectWithSha:... I broke the process
into 2 smaller methods, similar to the appraoch used in Grit.  We now
have:

- (NSData *)unpackObjectAtOffset:(NSUInteger)offset type:(GITObjectType*)objectType error:(NSError**)error;

This starts the process of unpacking any object at a given offset.  If it is
a normal object, we unpack it, and return.  If it is a delta object, we call:

- (NSData *)unpackDeltifiedObjectAtOffset:(NSUInteger)offset objectOffset:(NSUInteger)objOffset type:(GITObjectType)type error:(NSError**)error;

This reads the dela header, unpacks the base object (calls -unpackObjectAtOffset:..) , applies
the delta patch and returns the patched object.

The original loadObjectWithSha: loader is reduced to looking up the offset and starting the
unpack process.  This is still the only public facing method.