Bug fix in packSHA1() for cross-platform compatibility.
It seems that using NSString -characterAtIndex: in packSHA1() introduced some type of byte-order
dependency.  Perhaps it was due to the strchr() call to search the hexchars() string
and then doing a bit operation on the result?  I'm honestly not sure.  To fix the problem,
we convert the string to an NSData object with ASCII encoding, and then access the
bytes directly from a pointer.

In the process of debugging this, I also realized that packSHA1 returns a string
with ISOLatin1 encoding. Again, I am not exactly sure what it going on. There are
some sanity tests in GITUtilityBeltTests that explicitly test for a few cases
to demonstrate this 'feature'. Mostly these were for my own sanity, but they might
be useful at some point...

These changes make tests pass on i386 and ppc platforms.

I also reorganized unpackFromData / unpackFromBytes to do all of the work
with pointers in the latter function. In hindsight this was not necessary,
but it seems to make sense, should be more efficient and is not really any
less difficult to understand.