Change Git.framework install_name to @rpath, specify Run paths in executables.
@rpath is the most flexible option for a framework, which allows it to be
easily embedded or not.

Apple Documentation is here: http://tinyurl.com/ccxz9z

For the framework, we set the install_name (Installation Path) to be: @rpath

The UnitTests.app and the cli tools each link to the framework and specify
an rpath (Run path), which will be searched at runtime.  Libraries, ect..
found in the run path will be loaded, and dyld will substitute the
path for the @rpath variable allowing the lib to be loaded.

UnitTests.app embeds the framework in its private 'Frameworks' dir. Thus, we
add a '@executable_path/../Frameworks' run path to the target.

The cli apps current set a run path of $(CONFIGURATION_BUILD_DIR), which is
a special xcode variable that will resolve to 'build/<Config name>' ect.
This works assuming that the executables are installed in the same top-level
build dir as Git.framework.  This should work file for development purposes.