Refactor index lookup methods to use a binary search algorithm for sha1 lookups.
Start cleaning up the interface by adding -indexOfSha1: method that
finds the index of a sha1 in the sha1(v2) or offset+sha1(v1) table
and either returns the index or NSNotFound.  This is the same concept
as NSArray#indexOfObject: The main difference is that right now
indexOfSha1: is intended to be a private method, since other classes
should not really care about the 'index' in the table.

I think index lookups should be handled as follows:

Internal methods that require an index as a parameter should throw
an exception if the index is out of bounds, since it should never
be passed an illegal index unless there is a programming error.

The public methods for accessing properties associated with a sha1
(i.e. offset, CRC) just check for an NSNotFound and bail (optionally with an error).
Otherwise, the lookup succeeds.  If somehow -indexOfSha1: returns
a value outside of the index range, something bad has happened, and having
the exception will help track down the bug.