optimization[1/2]: faster routines for extracting data from core objects
When walking the commit graph, parsing essential information from commits can
be a potential bottleneck. This is not a big deal for small repositories, or
in cases where the entire graph not traversed. However, for large repos such
as git or the linux kernel, traversing the graph starting from HEAD takes
several seconds. Shark profiling shows that the NSScanner-based
GITCommit#parseRawData routine is slow, taking the same or more time than
(zlib) inflate. In constrast, zlib inflate is the major limiting step in
git-rev-list.

This change-set introduces some custom parsing code in an attempt to make this
go as fast as possible. The general approach is outlined below, with step 1
implemented in this change-set.

1. Remove NSScanner and drop straight down to raw C.
The slow steps seemed to be loading the scanner and allocating NSStrings. We
now access raw bytes directly and do not create an NSString unless we need it.
To drive the 'parser', define an objectRecord struct for each line or
substring that needs to be extracted. In some cases (i.e. sha1 strings) we
know the length, so specifying it explicitly saves some scanning. This code is
mostly used through a single-method Objective-C interface defined in a
"Parsing" category on GITObject.

2. Do _not_ parse the entire commit when the object is created.
Minimal nformation needed to traverse the graph:
+ commit-id (sha1) to uniquely identify the commit.
+ tree-id (sha1) to access the tree and object-ids associated with the commit.
+ list of parent-id (sha1)s.
+ raw committer date, ignoring the timezone.

Thus, for better efficiency, we extract only what we need, and cache the rest
in an NSData object for later extraction. This means that properties are now
fetched lazily:
+ author
+ authored
+ committer
+ committed
+ message

Currently lazy-fetching is as simple as possible, a call to any of the above
methods will cause all of the cached data to be parsed, filling in the rest of
the properties and releasing the cached data. This is simple to implement, and
the rationale is that long lists of commits/objects are likely to be needed
quickly during network transfer or pack file creation. If the information
associated with a commit is required, it is likely being displayed on demand
for a user who will not need to look at 18000 commits at once.

To summarize, it boils down to:
Extract structural information required for building a graph as fast as
possible. Fill in the rest when it is needed.

Signed-off-by: Geoff Garside <geoff@geoffgarside.co.uk>